requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
alembic==1.11.1
python-multipart==0.0.6
pandas==2.2.2
celery==5.3.1
redis==4.6.0
httpx==0.24.1
jinja2
python-dotenv==1.0.0
itsdangerous==2.1.2
aiofiles==23.1.0
requests>=2.31.0

